requirements.txt
captain
pydub
google-cloud-vision
google-cloud-storage
google-cloud-speech

